move realtime endpoint(s) to its own module
