+++ src/api/departures-grouped/index.ts
@@ -3,12 +3,10 @@ import { IDeparturesGroupedService } from '../../service/interface';
 import {
   DepartureGroupsPayload,
   DepartureGroupsQuery,
-  DepartureRealtimeQuery,
   DepartureFavoritesPayload,
   DepartureFavoritesQuery
 } from '../../service/types';
 import {
-  getDepartureRealtime,
   getDeparturesCursoredRequest,
   getDepartureFavoritesCursoredRequest
 } from './schema';
@@ -41,20 +39,6 @@ export default (server: Hapi.Server) => (service: IDeparturesGroupedService) => 
       const query = request.query as unknown as DepartureFavoritesQuery;
 
       return (await service.getDeparturesFavorites(location, query)).unwrap();
-    }
-  });
-  server.route({
-    method: 'GET',
-    path: '/bff/v1/departures-realtime',
-    options: {
-      tags: ['api', 'stops', 'departures', 'realtime'],
-      validate: getDepartureRealtime,
-      description:
-        'Get updated realtime information of all lines and quays passed as data'
-    },
-    handler: async (request, h) => {
-      const query = request.query as unknown as DepartureRealtimeQuery;
-      return (await service.getDepartureRealtime(query)).unwrap();
     }
   });
 };
