+++ src/api/departures-grouped/schema.ts
@@ -63,11 +63,3 @@ export const getDepartureFavoritesCursoredRequest = {
     cursor: Joi.string()
   })
 };
-
-export const getDepartureRealtime = {
-  query: Joi.object({
-    quayIds: Joi.array().items(Joi.string()).default([]).single(),
-    startTime: Joi.date(),
-    limit: Joi.number().default(5)
-  })
-};
