+++ src/api/realtime/schema.ts
@@ -1,0 +1,9 @@
+import Joi from 'joi';
+
+export const getDepartureRealtime = {
+  query: Joi.object({
+    quayIds: Joi.array().items(Joi.string()).default([]).single(),
+    startTime: Joi.date(),
+    limit: Joi.number().default(5)
+  })
+};
