+++ src/service/impl/departures-grouped/index.ts
@@ -1,10 +1,8 @@
 import { IDeparturesGroupedService } from '../../interface';
-import { DepartureRealtimeQuery } from '../../types';
 import {
   getDeparturesGrouped,
   getDeparturesGroupedNearest
 } from './departure-group';
-import { getRealtimeDepartureTime } from './departure-time';
 import { getDepartureFavorites } from './departure-favorites';
 
 export default (): IDeparturesGroupedService => {
@@ -18,9 +16,6 @@ export default (): IDeparturesGroupedService => {
             query,
             payload.favorites
           );
-    },
-    async getDepartureRealtime(query: DepartureRealtimeQuery) {
-      return getRealtimeDepartureTime(query);
     },
     async getDeparturesFavorites(payload, query) {
       return getDepartureFavorites(query, payload.favorites);
