+++ src/service/impl/realtime/journey-gql/departure-time.graphql-gen.ts
